+++ src/behaviors/MessageBehavior.php
@@ -3,14 +3,12 @@
 namespace Ryssbowh\CraftEmails\behaviors;
 
 use Ryssbowh\CraftEmails\Emails;
-use Ryssbowh\CraftEmails\helpers\RedactorHelper;
 use Ryssbowh\CraftEmails\models\Email;
 use craft\ckeditor\Field;
+use craft\ckeditor\Plugin;
 use craft\helpers\Html;
 use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
-use craft\redactor\assets\field\FieldAsset;
-use craft\redactor\assets\redactor\RedactorAsset;
 use craft\validators\HandleValidator;
 use yii\base\Behavior;
 use yii\helpers\Json;
@@ -38,40 +36,22 @@ class MessageBehavior extends Behavior
     }
 
     /**
-     * Get redactor input
-     *
-     * @param  ?string $redactorConfig
-     * @return string
-     */
-    public function redactorInput(?string $redactorConfig)
-    {
-        \Craft::$app->view->registerAssetBundle(FieldAsset::class);
-        RedactorAsset::registerTranslations(\Craft::$app->view);
-        $settings = RedactorHelper::getRedactorSettings($redactorConfig);
-        \Craft::$app->view->registerJs('new Craft.RedactorInput(' . Json::encode($settings) . ');');
-        $textarea = Html::textarea('body', $this->_parsedBody(), [
-            'id' => 'field-body',
-            'autocomplete' => 'off',
-            'style' => ['display' => 'none'],
-        ]);
-
-        return Html::tag('div', $textarea, [
-            'class' => [
-                'redactor',
-                'normal'
-            ],
-        ]);
-    }
-
-    /**
      * Get ckeditor input
      *
      * @since 2.1.0
      * @param  string $ckeConfig
      * @return string
      */
-    public function ckeditorInput(string $ckeConfig): string
+    public function ckeditorInput(?string $ckeConfig): string
     {
+        if (!\Craft::$app->plugins->isPluginEnabled('ckeditor')) {
+            return '<p class="error">' . \Craft::t('emails', 'You must install ckeditor in the settings') . '</p>';
+        }
+        try {
+            Plugin::getInstance()->ckeConfigs->getByUid($ckeConfig);
+        } catch (\Exception $e) {
+            return '<p class="error">' . \Craft::t('emails', 'The ckeditor config is not valid for this email') . '</p>';
+        }
         $config = [
             'type' => Field::class,
             'name' => 'Body',
@@ -79,30 +59,6 @@ class MessageBehavior extends Behavior
             'ckeConfig' => $ckeConfig
         ];
         $field = \Craft::$app->fields->createField($config);
-        return $field->getInputHtml($this->owner->body);
-    }
-
-    public function _parsedBody()
-    {
-        if (!StringHelper::contains($this->owner->body, '{')) {
-            return $this->owner->body;
-        }
-        return preg_replace_callback('/(href=|src=)([\'"])(\{([\w\\\\]+\:\d+(?:@\d+)?\:(?:transform\:)?' . HandleValidator::$handlePattern . ')(?:\|\|[^\}]+)?\})(?:\?([^\'"#]*))?(#[^\'"#]+)?\2/', function ($matches) {
-            /** @var Element|null $element */
-            list($fullMatch, $attr, $q, $refTag, $ref, $query, $fragment) = array_pad($matches, 7, null);
-            $parsed = \Craft::$app->getElements()->parseRefs($refTag);
-            // If the ref tag couldn't be parsed, leave it alone
-            if ($parsed === $refTag) {
-                return $fullMatch;
-            }
-            if ($query) {
-                // Decode any HTML entities, e.g. &amp;
-                $query = Html::decode($query);
-                if (mb_strpos($parsed, $query) !== false) {
-                    $parsed = UrlHelper::urlWithParams($parsed, $query);
-                }
-            }
-            return $attr . $q . $parsed . ($fragment ?? '') . '#' . $ref . $q;
-        }, $this->owner->body);
+        return $field->getInputHtml($this->owner->body, null);
     }
 }