+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Web Puzzlers Emails Changelog
 
+## 3.0.0 - unreleased
+
+> **Warning**
+> The support for Redactor has been dropped, please migrate to Ckeditor: Install the [plugin](https://plugins.craftcms.com/ckeditor) and choose a ckeditor config for each email
+
+### Changed
+
+- Craft 5 support
+- Dropped support for Redactor, Ckeditor is now the only wysiwyg editor supported
+
 ## 2.1.0 - 2023-07-21
 
 > **Warning**